+++ Plugins/Creature/Creature.hpp
@@ -136,6 +136,8 @@ private:
     ArgumentStack DoPerceptionUpdateOnCreature  (ArgumentStack&& args);
     ArgumentStack GetPersonalSpace              (ArgumentStack&& args);
     ArgumentStack SetPersonalSpace              (ArgumentStack&& args);
+    ArgumentStack GetCreaturePersonalSpace      (ArgumentStack&& args);
+    ArgumentStack SetCreaturePersonalSpace      (ArgumentStack&& args);
 
     CNWSCreature *creature(ArgumentStack& args);
     std::unordered_map<uint8_t, std::unordered_map<ObjectID, int16_t>> m_RollModifier;